pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dndcs"
version = "0.1.0"
description = "Modular backend for DnD character sheets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["click>=8.1", "pydantic>=2.6", "PyYAML>=6.0"]

[project.optional-dependencies]
ui = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "coverage>=7.5",
  "black>=24.4",
  "ruff>=0.5",
  "mypy>=1.10",
  "httpx>=0.27",
]

[project.scripts]
dndcs = "dndcs.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [".", "main-Core"]
include = [
  "dndcs",
  "dndcs.*",
  "dndcs_core",
  "dndcs_core.*",
  "dndcs_rulesets",
  "dndcs_rulesets.*",
  "webui",
  "webui.*",
]

[tool.setuptools.package-data]
"dndcs_rulesets" = [
  "fivee_stock/manifest.yaml",
  "fivee_stock/assets/**/*",
]
"webui" = ["static/**/*"]

[tool.coverage.run]
branch = true
source = ["dndcs", "main-Core/dndcs_core", "webui"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
pythonpath = ["."]
